Play/pause, forward, prev working

+++ LED Cube/src/com/schooler/ledcube/function/TimeFunction.java
@@ -2,6 +2,6 @@ package com.schooler.ledcube.function;
 
 public interface TimeFunction {
 
-	public void setTime(long millis);
+	public void setTime(double millis);
 	
 }
